New DAO v0.2

+++ src/main/java/entitylayer/Item.java
@@ -6,9 +6,9 @@ package entitylayer;
 public class Item {
     private int id;
     private String name;
-    private double price;
+    private int price;
 
-    public Item(int id, String name, double price) {
+    public Item(int id, String name, int price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -22,7 +22,7 @@ public class Item {
         return name;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 }
